backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
motor==3.6.0
beanie==1.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.1
openai==1.57.4
deepgram-sdk==3.8.0
boto3==1.35.83
weasyprint==62.3
python-multipart==0.0.19
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
mongomock-motor==0.0.34
